primera iteracion completa
