+++ angular-introduction/src/app/album-component/album-component.component.html
@@ -1,0 +1,12 @@
+<p>
+  album-component works!
+</p>
+
+<div class= "container">
+  <img src="{{ img1 }}" width="300" height="200" alt="" srcset="">
+
+  <img src="{{ img2 }}" width="300" height="200" alt="" srcset="">
+
+  <img src="{{ img3 }}" width="300" height="200" alt="" srcset="">
+
+</div>
