+++ angular-introduction/src/app/app.module.ts
@@ -1,0 +1,18 @@
+import { BrowserModule } from '@angular/platform-browser';
+import { NgModule } from '@angular/core';
+
+import { AppComponent } from './app.component';
+import { AlbumComponentComponent } from './album-component/album-component.component';
+
+@NgModule({
+  declarations: [
+    AppComponent,
+    AlbumComponentComponent
+  ],
+  imports: [
+    BrowserModule
+  ],
+  providers: [],
+  bootstrap: [AppComponent]
+})
+export class AppModule { }
