+++ angular-introduction/src/favicon.ico
