+++ angular-introduction/src/styles.css
@@ -1,0 +1,9 @@
+/* You can add global styles to this file, and also import other style files */
+
+.container{
+  display: flex;
+  justify-content: space-around;
+}
+p{
+  text-align: center
+}
