--- starter-code/iteration1/e2e/app.e2e-spec.ts
@@ -1,14 +1,0 @@
-import { LabAngularIntroductionPage } from './app.po';
-
-describe('lab-angular-introduction App', function() {
-  let page: LabAngularIntroductionPage;
-
-  beforeEach(() => {
-    page = new LabAngularIntroductionPage();
-  });
-
-  it('should display message saying app works', () => {
-    page.navigateTo();
-    expect(page.getParagraphText()).toEqual('app works!');
-  });
-});
